refactor(dashboard): extract DashboardInternals + fix sidebar open-folder no-op
- Extract SearchBar and DashboardLoadingState into DashboardInternals.tsx
  to bring DashboardPage.tsx from 316→271 lines (AC-12: ≤300)
- Fix DashboardSidebar open-folder button: no-op → invoke('dialog:folder:open')
